requirements.txt
streamlit
pandas
numpy
scikit-learn
joblib
requests
python-docx
pdfplumber
PyPDF2
spacy
sentence-transformers
rapidfuzz
python-dotenv
openai
pymongo
dnspython
google-generativeai
